Fix 12 bugs found during codebase audit
Critical fixes:
- predict.py: current_timestamp_ms=0 treated as None (falsy `or` → explicit None check)
- predict.py: remove dead `end_ms` variable, use integer division for mid_ms
- local_client.py: multi-shard results not sorted when decay_lambda=0 (always call apply_decay)
- async_client.py: missing `if state.scene_id:` guard in insert_batch causal linking
- async_client.py: predict_and_retrieve missing new API (PredictRetrieveResult, current_position, novelty)
- client.py + async_client.py: _known_collections empty for read-only clients (add _discover_collections)
- client.py + async_client.py + local_client.py: _find_latest_predecessor returns oldest instead of latest
- client.py + async_client.py + local_client.py: _list_active_epochs returns [0] instead of [] when empty
- memory.py: scroll sort was descending, should be ascending to match Qdrant default
- vjepa2.py: normalization mixed grid indices with scene bounds (now uses grid_shape)
- generic.py: unnecessary float64 conversion before tolist()
- hilbert.py: overlap expansion had no effect for zero-width (point) queries

New tests:
- PredictRetrieveResult defaults, novelty scoring, timestamp_ms=0 handling
- get_causal_context: window filtering, scene_id isolation, ordering, edge cases

+++ loci/adapters/generic.py
@@ -59,7 +59,7 @@ class GenericAdapter:
             y=y,
             z=z,
             timestamp_ms=timestamp_ms,
-            vector=embedding.astype(np.float64).tolist(),
+            vector=embedding.tolist(),
             scene_id=scene_id,
             scale_level=scale_level,
             confidence=confidence,
